pyproject.toml
[tool.poetry]
name = "aiotraq-monorepo"
version = "0.1.2"
description = "Async ready traQ Developper toolkit"
authors = ["toshi00"]
license = "MIT"
readme = "README.md"
repository = "https://github.com/toshi-pono/aiotraq"
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"


[tool.poetry.group.api.dependencies]
openapi-python-client = "^0.19.1"


[tool.poetry.group.dev.dependencies]
aiotraq = { path = "libs/aiotraq", develop = true }
aiotraq-message = { path = "libs/message", develop = true }
python-dotenv = "^1.0.1"
matplotlib = "^3.9.0"
aiotraq-bot = { path = "libs/bot", develop = true }
ruff = ">=0.2,<0.4"
mypy = "^1.10.0"
pillow = "^10.3.0"

[tool.mypy]
show_error_context = true
show_column_numbers = true
ignore_missing_imports = true
disallow_untyped_defs = true
no_implicit_optional = true
warn_return_any = true
check_untyped_defs = true
exclude = "libs/*"


[tool.ruff]
exclude = [".venv", "venv", "__pycache__", ".git"]

line-length = 120
indent-width = 4

[tool.ruff.format]
quote-style = "double"

[tool.ruff.lint.mccabe]
max-complexity = 10
